refactor(analysis): split content read helpers
