+++ docs/architecture-consolidation-plan.md
@@ -60,7 +60,7 @@ fixtures:
 | Area | Current file | Approx. lines | Owner direction |
 | --- | --- | ---: | --- |
 | Content complexity | `crates/tokmd-analysis/src/content/complexity.rs` and `crates/tokmd-analysis/src/content/complexity/` | `tests/unit.rs` 10; cyclomatic test owner 562; cognitive test owner 194; function test owner 404; nesting test owner 177; production owner modules <=187 | Scoring, nesting, and function-span helpers now live under owner modules; function-detection, cyclomatic, cognitive, and nesting proof tests now have dedicated owners; remaining work is mostly aggregation cleanup |
-| Content I/O helpers | `crates/tokmd-analysis/src/content/io.rs` and `crates/tokmd-analysis/src/content/io/` | coordinator 336; byte owner/tests 108; tag owner/tests 82 | Keep bounded file reads in `io.rs`; keep text detection, BLAKE3 hashing, capped file hashing, and entropy in `io/bytes.rs`; keep tag counting and delimited matching in `io/tags.rs`. Preserve `crate::content::io::*` call sites and the `analysis_content_assets` proof scope. |
+| Content I/O helpers | `crates/tokmd-analysis/src/content/io.rs` and `crates/tokmd-analysis/src/content/io/` | coordinator 68; read owner/tests 267; byte owner/tests 108; tag owner/tests 82 | Keep `io.rs` as the public facade; keep bounded file reads in `io/read.rs`; keep text detection, BLAKE3 hashing, capped file hashing, and entropy in `io/bytes.rs`; keep tag counting and delimited matching in `io/tags.rs`. Preserve `crate::content::io::*` call sites and the `analysis_content_assets` proof scope. |
 | Analysis orchestration | `crates/tokmd-analysis/src/analysis/mod.rs` and `crates/tokmd-analysis/src/analysis/` | coordinator 152; setup 20; files 53; outputs 22; enricher owners <=160 | Derived/source setup, file collection, optional output storage, and inventory/content/git/semantic/code-quality/effort enrichers now live in owner modules while `analyze` remains the public coordinator. Keep `analysis_orchestration` proof scoped to both the historical `analysis.rs` path and the new `analysis/**` owner paths during the file-to-folder transition |
 | Analysis git report | `crates/tokmd-analysis/src/git/mod.rs`, `git/freshness.rs`, and `git/churn.rs` | coordinator 221; freshness/code-age owner 192; predictive churn owner/tests 174 | Keep export-row matching, hotspot, bus-factor, coupling, and intent assembly in `mod.rs` for now; keep freshness, age buckets, and refresh trend in `freshness.rs`; keep predictive churn in `churn.rs`. Preserve the `analysis_git` proof scope. |
 | Analysis archetype inference | `crates/tokmd-analysis/src/archetype/mod.rs` and `crates/tokmd-analysis/src/archetype/rules.rs` | coordinator 74; rule owner/tests 403 | Keep export-row filtering and path normalization in `mod.rs`; keep ordered Rust/Next.js/container/IaC/Python/Node archetype rules and rule tests in `rules.rs`. Preserve `analysis_security_identity` proof coverage. |
